requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
black==22.3.0
click==8.1.0
Django==4.0.3
django-crispy-forms==1.14.0
django-osm-field==0.5.0
flake8==4.0.1
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.8
pytest==7.1.1
pytest-django==4.5.2
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
python_version==3.8.10
